+++ CHANGELOG.md
@@ -40,6 +40,8 @@ release date when you use `npm version` (see `README.md`).
 
 - Allow for large contents of JSON in details panel
 
+- Added Content Security Policy - the same one as on DeltaChat Desktop.
+
 ## [0.12.1][] - 2022-06-29
 
 ### Added
@@ -292,7 +294,5 @@ release date when you use `npm version` (see `README.md`).
 [0.12.1]: https://github.com/webxdc/webxdc-dev/tree/v0.12.1
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.13.0...HEAD
 [0.13.0]: https://github.com/webxdc/webxdc-dev/tree/v0.13.0
-
-
-[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.13.1...HEAD
+[unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.13.1...HEAD
 [0.13.1]: https://github.com/webxdc/webxdc-dev/tree/v0.13.1